Move scripts package to src and rename to trail_route_ai
