+++ src/trail_route_ai/planner_utils.py
