+++ tests/conftest.py
@@ -1,0 +1,5 @@
+import os
+import sys
+
+# Ensure the src package is importable
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src')))
